html

+++ src/main/java/team/placeholder/internalprojectsmanagementsystem/service/project/ProjectService.java
@@ -9,7 +9,7 @@ public interface ProjectService {
     ProjectDto save(ProjectDto projectDto);
 
     List<ProjectDto> getAllProjects();
-    ProjectDto getProjectById(int id);
+    ProjectDto getProjectById(long id);
     ProjectDto getProjectByName(String name);
 
     ProjectDto updateProject(ProjectDto projectDto);
